Restructure project to be cleaner
