+++ opal/module/__init__.py
@@ -1,0 +1,4 @@
+from opal.module.neu_mf import NeuMF
+from opal.module.neu_mf_module import NeuMFModule
+
+__all__ = ['NeuMF', 'NeuMFModule']
