--- opal/score/collaborative_filtering/__init__.py
@@ -1,3 +1,0 @@
-from .neu_mf import NeuMF
-
-__all__ = ['NeuMF']
